The gnuplot loss script to create a chart.
As Hartwig mentioned the script for generating a chart of the number of lost packets. If you would like to increase the size please change the format in line 12. From practical use of the chart script it is a good idea to start the test ~3 minutes after you've started the udpping an leave afterwords time to avoid bouncing the graph into the agenda.